+++ custom_components/pweb_amano/const.py
@@ -4,3 +4,5 @@ from datetime import timedelta
 DOMAIN = "pweb_amano"
 
 DEFAULT_SCAN_INTERVAL = timedelta(minutes=30)
+
+SERVICE_REGISTER_DISCOUNT = "register_discount"